[Step1] Frame에 템플릿 메소드 패턴 적용
