+++ src/main/java/domain/frame/NormalFrame.java
@@ -1,0 +1,17 @@
+package domain.frame;
+
+import domain.PhaseResult;
+import domain.phase.NormalFramePhaseCollection;
+
+public class NormalFrame extends Frame {
+
+	public NormalFrame() {
+		super(new NormalFramePhaseCollection());
+	}
+
+	@Override
+	public PhaseResult getPhaseResult() {
+		return phaseCollection.getFrameResult(remainBowlingPins);
+	}
+
+}
